refactor: parallelize test suite with xdist loadgroup + session-scoped pools
- Session-scoped fixtures (pool, pip_pool, apt_pool, env_pool, torch_pool)
  replace per-module pools so each pool type is provisioned only once
- Tests grouped by pool type via @pytest.mark.xdist_group for concurrent
  execution across xdist workers
- Container provider uses ephemeral ports (-p 0:22) to avoid port conflicts
  when multiple pools start concurrently
- timeout_func_only=true prevents fixture setup from counting against test
  timeouts
- Restructured test files from nested dirs to flat layout with xdist markers
- Fix vcpus type to float in actor messages
- Add pyright type narrowing for instance_info() in tests

+++ skyward/actors/messages.py
@@ -53,7 +53,7 @@ class InstanceMetadata:
     instance_type: str = ""
     gpu_count: int = 0
     gpu_model: str = ""
-    vcpus: int = 0
+    vcpus: float = 0
     memory_gb: float = 0.0
     gpu_vram_gb: int = 0
     region: str = ""
@@ -180,7 +180,7 @@ class InstanceRunning:
     instance_type: str = ""
     gpu_count: int = 0
     gpu_model: str = ""
-    vcpus: int = 0
+    vcpus: float = 0
     memory_gb: float = 0.0
     gpu_vram_gb: int = 0
     region: str = ""
